Updated ThemeToggle component
