+++ src/Components/Input/Input.module.css
@@ -31,7 +31,7 @@
   border: 2px solid var(--input-border-color);
   border-radius: 5px;
   cursor: pointer;
-  transition: transform 0.2s ease, background-color 0.2s;
+  transition: transform 0.2s ease, background-color 0.2s, border 0.2s;
 }
 
 .input:focus {
