+++ src/Components/Navigation/Navigation.jsx
@@ -1,10 +1,6 @@
 import styles from './Navigation.module.css';
-import { useContext, useState } from 'react';
-import { ThemeContext } from '../../context/ThemeProvider';
-import { ReactComponent as Sun } from '../../assets/sun.svg';
-import { ReactComponent as Moon } from '../../assets/moon.svg';
-import Logo from '../Logo/Logo';
-import Tooltip from '../Tooltip/Tooltip';
+import ThemeToggle from './ThemeToggle';
+import Logo from './Logo';
 
 export default function Navigation() {
   return (
@@ -16,34 +12,3 @@ export default function Navigation() {
     </header>
   );
 }
-
-const ThemeToggle = () => {
-  const { theme, toggleTheme } = useContext(ThemeContext);
-  const [showTooltip, setShowTooltip] = useState(false);
-  const switchPosition = theme === 'dark' ? styles.left : styles.right;
-  const handleShowTooltip = () => setShowTooltip(true);
-  const handleHideTooltip = () => setShowTooltip(false);
-
-  return (
-    <div
-      className={styles['theme-container']}
-      onMouseEnter={handleShowTooltip}
-      onMouseLeave={handleHideTooltip}
-    >
-      <span className={`${styles.switch} ${switchPosition}`}>
-        {theme === 'light' && (
-          <Sun
-            className={`${styles.sun} ${styles['theme-svg']}`}
-            onClick={toggleTheme}
-          />
-        )}
-        {theme === 'dark' && (
-          <Moon
-            className={`${styles.moon} ${styles['theme-svg']}`}
-            onClick={toggleTheme}
-          />
-        )}
-      </span>
-    </div>
-  );
-};
