+++ src/Components/UI/Card.module.css
@@ -7,7 +7,7 @@
   display: grid;
   background-color: var(--card-bgcolor);
   box-shadow: 0 0 5px 1.75px var(--card-shadow);
-  transition: background-color 0.2s;
+  transition: background-color 0.2s, box-shadow 0.2s;
 }
 
 @media (min-width: 450px) {
